hdfs_consumer/requirements.txt
confluent-kafka
python-dotenv
pyspark==3.5.0
py4j==0.10.9.7
